fix: do not export SnapAccountService* + move them to types.ts
