+++ packages/accounts-controller/src/index.ts
@@ -15,9 +15,6 @@ export type {
   AccountsControllerAccountBalancesUpdatesEvent,
   AccountsControllerAccountTransactionsUpdatedEvent,
   AccountsControllerAccountAssetListUpdatedEvent,
-  SnapAccountServiceAccountBalancesUpdatedEvent,
-  SnapAccountServiceAccountTransactionsUpdatedEvent,
-  SnapAccountServiceAccountAssetListUpdatedEvent,
   AllowedEvents,
   AccountsControllerEvents,
   AccountsControllerMessenger,
